refactor(ghgi): move inventory page logic to shared components
Extract GHGI inventory page implementations into components/GHGI/inventory-pages and point city and legacy routes at the shared modules. This removes city route dependencies on app/[lng]/[inventory] page files while preserving existing behavior.
