+++ app/src/app/[lng]/cities/[cityId]/GHGI/[inventory]/data/[step]/page.tsx
@@ -1,11 +1,1 @@
-"use client";
-
-import AddDataSteps from "@/app/[lng]/[inventory]/data/[step]/page";
-
-/**
- * JN route wrapper for add data steps page
- * Reuses the GHGI add data steps page component to keep DRY
- */
-export default function GHGIAddDataSteps() {
-  return <AddDataSteps />;
-}
+export { default } from "@/components/GHGI/inventory-pages/add-data-steps-page";
