+++ app/src/app/[lng]/cities/[cityId]/GHGI/[inventory]/data/manage-sectors/page.tsx
@@ -1,11 +1,1 @@
-"use client";
-
-import ManageSubSectors from "@/app/[lng]/[inventory]/data/manage-sectors/page";
-
-/**
- * JN route wrapper for manage sectors page
- * Reuses the GHGI manage sectors page component to keep DRY
- */
-export default function GHGIManageSubSectors() {
-  return <ManageSubSectors />;
-}
+export { default } from "@/components/GHGI/inventory-pages/manage-sub-sectors-page";
