+++ app/src/app/[lng]/cities/[cityId]/GHGI/[inventory]/data/page.tsx
@@ -1,11 +1,1 @@
-"use client";
-
-import AddDataIntro from "@/app/[lng]/[inventory]/data/page";
-
-/**
- * JN route wrapper for data intro page
- * Reuses the GHGI data intro page component to keep DRY
- */
-export default function GHGIAddDataIntro() {
-  return <AddDataIntro />;
-}
+export { default } from "@/components/GHGI/inventory-pages/add-data-intro-page";
