+++ app/src/app/[lng]/cities/[cityId]/GHGI/[inventory]/data/review/page.tsx
@@ -1,11 +1,1 @@
-"use client";
-
-import ReviewPage from "@/app/[lng]/[inventory]/data/review/page";
-
-/**
- * JN route wrapper for review page
- * Reuses the GHGI review page component to keep DRY
- */
-export default function GHGIReviewPage() {
-  return <ReviewPage />;
-}
+export { default } from "@/components/GHGI/inventory-pages/review-page";
