+++ app/src/app/[lng]/cities/[cityId]/GHGI/[inventory]/settings/page.tsx
@@ -1,7 +1,1 @@
-"use client";
-
-import SettingsPage from "@/app/[lng]/[inventory]/settings/page";
-
-export default function CityInventorySettingsPage() {
-  return <SettingsPage />;
-}
+export { default } from "@/components/GHGI/inventory-pages/settings-page";
